requirements.txt
onnx_graphsurgeon 
opencv-python
imread-from-url
onnxruntime-gpu